Moved skill.ts outside its folder as it was the only file
