+++ app/certificates/components/CertificatesView.tsx
@@ -20,7 +20,7 @@ import SearchInput from "@/components/Inputs/SearchInput";
 import CertificatesList from "@/components/MaterialLists/CertificatesList";
 import { Button } from "@/components/shadcn/ui/button";
 import certificateDatabase from "@/database/certificates";
-import skillDatabase from "@/database/skills/skills";
+import skillDatabase from "@/database/skills";
 import CertificateSlugEnum from "@/enums/MaterialSlugEnums/CertificateSlugEnum";
 import SkillSlugEnum from "@/enums/SkillSlugEnum";
 import SkillTypesEnum from "@/enums/SkillTypesEnum";
