+++ app/skills/page.tsx
@@ -2,7 +2,7 @@ import HeadingOne from "@/components/Text/HeadingOne";
 import PageDescription from "@/components/UI/PageDescription";
 import developerName from "@/constants/developerName";
 import SkillList from "../../components/MaterialLists/SkillList";
-import { skillKeys } from "@/database/skills/skills";
+import { skillKeys } from "@/database/skills";
 
 const description = `
 	Explore my collection of skills on various topics.
